AG: Check for device when removing CLCC message.
When removing the CLCC response time out message,
check for the device which it belongs to. Removing
all the messages leading to not sending CLCC response
to one of the headset requested it when two connected
headsets sent AT+CLCC simultaneously.

CRs-Fixed: 988211
Change-Id: I21ecbf78dca5c14a915a7fa7d13da172371c5c5e

Get last dialled number for IMS and WIFI calls

When querying for last dialled number from Telephony,
include IMS and WIFI calls.

CRs-Fixed: 1097647
Change-Id: I8b3a951fd3ce4f576968e4732f18a85dbe2251f4

Retry once more if outgoing HF connection fails

When first outgoing HF connection fails for some
reason, attempt HF connection again after 2500msec

CRs-Fixed: 2000407
Change-Id: I4168dc2c08f7fa74dd42004cc92ba78dfc65a2df

HFP AG: Fix the condition to check if SCO is connected

Two headsets hs1, hs2 are connected and SCO is created with
hs2. If hs1 is disconnecting, we move to MultiHFPending state
from AudioOn state and then move back to AudioOn state. We
should check for audio state and SCO device instead of
checking if we are in AudioOn state.

CRs-Fixed: 968433
Change-Id: Ic48f4403b544ee11258d47807630a435f736d8f9

HF: Do not remove connect defer messages in Connected state

When we are in AudioOn state and  headset2 is connected,
we disconnect SCO and move to Connected state. If we remove the
connect defer message from the queue, we will not be able to
connect to the headset2.

CRs-fixed: 916103
Change-Id: Ideb077108ab6b7e7a0caf0f881960c7c53bbd330

Disconnect SCO for waiting call for blacklisted devices

Some remote devices are misbehaving when there is an active
call and MT call is in progress. We send call indicators for
active call and MT Call in progess. When active call is
terminated, we send call end indicator only to remote. This
is confusing remote and remotes are going into bad state.
Blacklist remote devices to disconnect SCO when active call is
ended, fake MT call indicator again with some delay.

CRs-Fixed: 2019234
Change-Id: Ie69b99e88d26886eb0b362f553b67c52f93f5b47

Don't allow SCO when A2DP playback is going on.

When A2DP playback is going on, don't allow SCO
to be created.

CRs-Fixed: 2038052
Change-Id: I0aa3a7834d4d41ef5e9124d6ddf193951966736d