OPP: Update batch status when sdp failed
Use case:
1) Pair with remote device
2) Turn off Bluetooth of the remote device
2) From the DUT send a file to device from the
   list of Paired device.

Failure:
Failure:
There is no error message poped up on the phone for failure of transfer.
Even after turning on the BT on the remote device resend the file still can
NOT send successfully unless Turn off and on the Bluetooth on the phone.

Reproducibility : 5/5

Fix:
Update the current batch status as failed when receive invalid sdp
record and send transfer complete event with status as failure.

Test: Checked if file sending is successful after remote device is turned back on
after initial failed transfer and user gets notificaiton of failure.

Bug: 37978462
CRs-Fixed: 2054433
Change-Id: I02581ee779624b17e7deedaaacaf91922c47f70c